+++ R/make_unique_id.R
@@ -11,13 +11,13 @@
 
 make_unique_id <- function(.data, ..., new_col = "hash") {
   check_polars_data(.data)
-  if (new_col %in% pl_colnames(.data)) {
+  if (new_col %in% names(.data)) {
     rlang::abort(
       paste0('Column "', new_col, '" already exists. Use a new name with the argument `new_col`.')
     )
   }
   vars <- tidyselect_dots(.data, ...)
-  if (length(vars) == 0) vars <- pl_colnames(.data)
+  if (length(vars) == 0) vars <- names(.data)
   out <- .data$with_columns(
     pl$struct(vars)$hash()$alias(new_col)
   )