+++ R/pivot_longer.R
@@ -36,7 +36,7 @@ pivot_longer.RPolarsDataFrame <- function(data, cols, ..., names_to = "name",
 
   check_polars_data(data)
 
-  data_names <- pl_colnames(data)
+  data_names <- names(data)
   value_vars <- tidyselect_named_arg(data, rlang::enquo(cols))
   id_vars <- data_names[!data_names %in% value_vars]
   out <- data$melt(